Move 'handleException' logic from 'RequestHandlerCache' to 'OutgoingAsync'.
